Integrate dprint (#871)
ref: https://github.com/pankona/hashira/pull/866, https://github.com/pankona/hashira/issues/869

This configuration is not my preference. It keeps zero configration as possible we can.
Because I respect the pankona wish as he said as https://github.com/mobu-of-the-world/mobu/pull/489#issuecomment-1172006167.

---

Ignore files around gitbook

Because some of them are generated by tools.
And we may drop gitbook dependency. See #870